Improve management of test(suite) directory names
With the move to python3, the TEST_DIR variable exposed
by context.py, based on sys.modules['__main__'].__file__,
now conveys an absolute path. This has detrimental consequences
on the production of STR qualification results as the
extra directory components end up in testcase identifiers.

This change robustifies the computation of tests' homedir
and reldir attributes to account for the __file__ change,
and replaces uses of the TEST_DIR variable by uses of the
reldir test attribute.

Part of V523-030 (gnatcov qualkit with e3/python3 on lima)

Change-Id: I3e7b075081bb805e014d2db099c6f751c4715492
Depends-On: Id7de2b90cd84c997cbbcb382798268caa1e85014