Add decision/Operands tests for quantified expressions
Test the use of quantified expressions used as decision
operand(s), showing that the tool properly distinguishes the outer
decision from the inner predicate.

Verify the correctness of results for
for all and for some, iterating over an array index (for I
in Arr'Range) or directly over array values (for X of Arr).

V309-022 (gnatcov qualification for quantified expressions)

Change-Id: Ic4e71ec426dadeceef403f03e760ec811ac55a9f