Traces_ELF: cope with odd values DW_AT_low_pc from recent binutils
For ARM targets, recent versions of binutils encode the ARM/Thumb
information setting the address LSB to 0 or 1. Our data structures need
to deal with the actual address: one problem is that this encoding
breaks the proper nesting invariant for address ranges.

This commit adds special handling for this encoding to unset the LSB,
and thus fix the invariant violation for ARM programs.

Change-Id: I597449917085fb7dd9465414554da088bb3d84ca
TN: V614-006