Add diff-based testing capabilities to gnatcov testsuite
To check that gnatcov produces the expected result, we usually verify
the xcov + report format against a specification of coverage expectations.
This scheme does not check metadata (such as SCO annotations, or messages)
and is unsuitable for testing such things.

This commit add support to check a xcov report file against a baseline.
Note that the work could easily be enhanced to support any kind of format.

Change-Id: I78fdab0160e3628e80d14d7455c9fc7727702476
TN: V426-009