Instrument-c: Fix compilation error
A call to an overloaded version of Append_SCO which will be
introduced in a later patch
(change-id I43fd7a9b729aacd00f9a01d6eb34b177e6f7ae48) is present
in gnatcov due to a patch splitting mistake.

Correct this while the above patch is not merged.

Change-Id: I00c5bd0b95f49f819fb1e79dbf227dc0f17e81d6
TN: V426-009