Workaround GNATemulator abnormal termination
If GNATemulator runs as a background process in an interactive shell,
the Linux kernel will send a SIGTTIN when GNATemulator tries to read the
standard input (which is the interactive shell by default). This will
result in an abnormal GNATemulator termination, and the test failing.
Redirecting the standard input to /dev/null works around this issue.

Change-Id: Icee92e82515ed29f60a8ee3d8294e03fc07322d2
TN: U915-014