annotations-dhtml.adb: always fill a non-empty project name
The Project_Name function can return an empty string if the given source
could not be attached to any project of the tree. As we always want to
fill the "project" field (otherwise, the report will be buggy), let's
fallback to a default of "Other sources" in that case.

This case can happen when "gnatcov coverage" loads a copy of the project
that was passed to "gnatcov instrument": as absolute names for source
files in the original project are recorded in SID files, "gnatcov
coverage" fails to relate these sources to the copied project, as its own
sources have different absolute names.

Change-Id: I18eb0a2cf9e1af1b9e7e5869208eceac6cf09734
TN: V627-017