Fix the computation of the configuration setup file basename
If we passed --runtime-project=<full_path> to ``gnatcov instrument``,
the name of the configuration file, used to load ``gnatcov instrument``
defaults, was not correctly computed, as it used that full path instead
of computing its basename.

Also, as we were lower casing this <full_path> when loading it with
GNATCOLL.Projects, it could fail if it had an upper case letter.

TN: V615-017
Change-Id: I1fa0b374873238ea9b30a06385616cfe482baf0c