C instrumentation: warn on parse inconsistencies
To keep information about macro expansion, we parse the unpreprocessed
view of the source file, and to instrument it, we parse the preprocessed
view (as we can't rewrite an unpreprocessed file). We used to error out if
there were inconsistencies between the two parses: we should be more
resilient as this will only result in a degraded report. Let's simply warn
the user.

Change-Id: Ie32f118d4fc62b0d8a0a7b1fb4607ec244da1f0d
TN: V728-002