Introduce a Undetermined_Coverage state and report category
This change introduces a new coverage state for SCOs with
undetermined coverage (e.g. corresponding to non-instrumented
code).

This also adds a new structure to track the set of SCOs for which
the corresponding code was not instrumented. These are now reported
as undetermined coverage instead of as violations.

SCOs with undetermined coverage are reported in a new report
section, distinct from non-coverable items, since here we
really do not know if the code asociated with these SCOs was
executed or not.

TN: V419-003

Change-Id: Ic9b167ae58461a188c291a8dc3a433491cf14de8