coverage runtime: remove std=c89 from C compiler flags
We want the code of the coverage runtime to be C89-compliant, to support
older toolchains. Nevertheless, system headers are not necessarily
compatible with C89 (in which case the toolchain would support a more
recent version of C). Remove the strict requirement std=c89 in
gnatcov_rts.gpr, as it prevents the compilation of the coverage runtime
in such cases,  and ensure C89 compatibility by other means (as a test
in gnatcov testsuite).

Change-Id: I9706b57651d6d530f944525953c6587580d4c812
TN: V808-022