Make codebase compilable with -gnat2022
Even if we don't have -gnat2022 in the compilation switches, it can be
useful to enable it when developping (to use the very convenient
generalized to non-scalar 'Image feature). Make the code Ada 2022
compliant, and qualify ambiguous calls to Ada.Containers.Vectors.Append.

As vectors are allowed to have aggregates in Ada 2022, calls to Append
become ambiguous as the Append overload accepting a vector as a second
argument becomes a possible choice when the aggregate is not qualified.

Change-Id: I2c3506d289f1ddb7dc9019bc7624e12924eb76d1
TN: V816-018