instrument-c*: fix and rename Insert_Text_After
Preparatory work for C++. This discusses an upcoming change wrapping
conditions in ternary expressions.

Just like what we did for Insert_Text_After_Start_Of /
Insert_Text_Before_Start_Of, clarify the terminology for
Insert_Text_After.

It used to insert text at the end of the cursor location, after any
previously inserted string at the same location. Change the second part
to insert text _before_ any previously inserted string at the same
location. This is usually what we want as we use this function when
inserting witness calls. The only case where we may insert text at the
same location is when we instrument a decision for MC/DC: the end of the
decision witness and the end of the last condition witness may be at the
same location. As we instrument the decision (at the top-level), before
the conditions (nested in the decision), we want to insert the condition
witness closing parenthesis / ternary expression wrapper, before the
closing parenthesis of the decision witness.

Without this fix:
a && b
-> witness_dec (a && b)
                      ^inserting ")" at the decision end sloc
-> witness_dec (witness_cond (a ? 1 : 0), witness_cond (b) ? 1 : 0)
                                                           ^inserting
                                                            after ")"

With this fix:
-> witness_dec (a && b)
-> witness_dec (witness_cond (a ? 1 : 0), witness_cond (b ? 1 : 0))
                                                          ^inserting
                                                          " ? 1 : 0)"
                                                          before ")"

Note that we need to wrap the decision in a ternary expression as well:
the order in which we will be inserting two identical strings thus won't
matter, but we are faithful to the intent by doing that (i.e. a closing
parenthesis should balance its starting parenthesis, and not another one).

The change takes the opportunity to fix glitches in the comments
associated with those other subprograms.

TN: V816-018
Change-Id: If4c8343f180d5436ac8421b4e1f730af4529a096