Add "not null" to predicate type def for QuantExpr tests
Access to predicate functions in Quantified Expression tests
are expected never to be null. Mention this explicitly in the
type definition, which will prevent access checks and avoid
possible CFG analysis warnings on the corresponding cond branch.

This cures a couple of spurious test failures in binary trace
mode for optimized runs on some targets, is perfectly legit in
the tests context and doesn't invalidate the original tests
purpose at all.

We have a separate ticket to investigate the failure mode
further.

Part of V822-011.

Change-Id: I72af20ce91af83cc60075d04ba50404c73c1fc04