Instrument: hierarchy of tagged types to group language operations
This is a pure refactoring (no behavior change). In order to avoid as
much as possible case statements to handle all instrumented languages in
language-agnostic code, introduce a hierarchy of tagged types that
provide language-specific operations.

This allows replacing case statements/duplicated code with dispatching
calls in instrument.adb, and clarifies the set of operations required in
order to support a language for instrumentation.

TN: V816-018
Change-Id: I1d870cddfbadb594f9709e1a0fdadef4f161954a