gnatcov_rts: constify the data to write in output callbacks
In C++, using gnatcov_rts's STR macro with a string literal triggers the
following:

  warning: ISO C++ forbids converting a string constant to 'char*'
  [-Wwrite-strings]
  note: in definition of macro 'STR'
   33 | #define STR(x) (struct gnatcov_rts_string) { x, sizeof (x) - 1 }

This is because the "str" field in the gnatcov_rts_string struct is
declared as "char *", whereas the string literal has type
"const char *".

All gnatcov_rts_string are supposed to be statically allocated, so there
is in fact no need to have a pointer to mutable chars. Move it to "const
char *" to remove the warning. This in turn requires the various output
routines in gnatcov_rts to take a pointer to constant data rather than a
pointer to mutable data, which is an improvement, as these routines are
not supposed to write to the data buffers, only to read from it.

As far as C is concerned, this is just a refactoring: no behavior
change expected.

TN: V816-018
Change-Id: If77379aca65a64659b97e1181691825841f4d896