Instrument: do not instrument C headers source files
Currently, the code in C headers is not instrumented, and yet the
instrumenter runs partially on C headers, creating both an instrumented
source for it in the project object directory, but also creating a
source for its (empty) coverage buffers.

Just skip the instrumentation for C headers file so that we do not
create these partially instrumented sources for it. No user-visible
behavior change, as we did not generate coverage obligations for these
partially instrumented source files.

TN: V816-018
Change-Id: Ifdaafc76443b1b8a0840324cde72ce53bb87039e