Introduce a SPARK compatibility mode
This mode ensures the instrumented code can be legal ghost code.

This is needed because we cannot determine if a generic will be
instantiated in a ghost code context or not, and thus
instrumentation cannot introduce code in generics that is invalid
in a ghost context.

This compatibility mode is not activated by default because it can
lead to witness calls being optimized out by the compiler, and thus
result in spurious coverage violations.

Change-Id: Ief50e928eb6f2c36f37e9b1fe4721139d5c97f9e
TN: V819-013