Instrument: enable main autodump for sources of disabled languages
For instance, in an Ada+C project with a C main, instrumenting for Ada
coverage only (--restricted-to-languages=Ada) will now insert the
automatic dump of coverage buffers in the C main, even though C is not
enabled for instrumentation (it was skipped before this commit).

This new behavior is what is probably the most convenient for users, and
is easy to overcome if unwanted: just remove the instrumented main in
the object subdirectory.

TN: V816-018
Change-Id: I50965b774fdbd980307857edfb6522b22c4644f1
Depends-On: I459756773beb3a78c67c8b85f63ffe4ce8dcac08