instrument-c.adb: refactor the emission of decls/definitions
Introduce helper subprograms to do the string concatenations and the
file writing/instrumenter insertions, and rewrite the manual string
formattings to use these.

After this change, empty lines are automatically added after each
declaration (we used to have most declarations on a single line, which
is a problem for debugging), and it is much easier when looking at the
instrumenter code to understand what is being emitted (an array? a
function? a simple variable/constant?).

This refactoring will also make it easier in the upcoming C++ support to
automatically append 'extern "C"' bits before each
definition/declaration.

TN: V816-018
Change-Id: I923c33f7634de2da1f9e414607216e50d1a1a432