instrument-c.adb: fix coverage buffer types
In instrumented C units, dummy extern declarations for coverage buffers
are currently declared as "unsigned char *" (the actual type for buffers
defined in buffer units is "uint8_t *", but since instrumented C units
do not perform direct access to them, and since uint8_t requires an
include, which may interfere with user code, this is fine).

However, the dummy extern declarations for the witness functions take
"void *" arguments. This commit harmonizes both to consistently use
"unsigned char *", and remove related casts. This is just a cleanup:
there was no known problem with this inconsistency.

TN: V816-018
Change-Id: I0ed7f6d0b333b60f8e8d81217125e3dbc5e41a6b