Instrument: change pattern for statement witness in expr context
Currently, when the instrumenter for C has to insert a call to a witness
function in an expression context, it assumes that the expression is a
decision and turns:

  EXPR

into:

  WITNESS (...) && EXPR

Upcoming work to support C++ (for instance the instrumentation of "while
(T VAR = EXPR)" constructs) will have to insert witness calls in
expressions that are not decisions, and thus where the && operator is
not applicable. Use the comma operator instead, which works in all known
cases. For the example above, this yields:

  (WITNESS (...), EXPR)

Change-Id: I5d37eea5f19efddbddd63a231c6d2c0d7def0696
TN: V816-018