Instrument: fix witness calls for decisions/conditions and comma exprs
It is legal in C to use the comma operator in contexts where gnatcov
expects a decision. For instance:

  if (a, b)

For such decisions, the C instrumenter is currently not inserting
witness calls correctly:

  if (witness (params, a, b))

The comma operator interfers with the comma for the list of arguments.
We should get instead:

  if (witness (params (a, b)))

This is true for witness calls for decisions, and for conditions. This
patch fixes this bug.

TN: V816-018
Change-Id: I6f7febc6dab18fba9e0738945922aae467a4bc6c