Instrumentation: allow the extension of the SCO_Unit_Table
Preparatory work for coverage analysis of C headers.

When traversing a body file, some SCOs may be attached to e.g. an
included file. In this case, we want to attach the SCOs to the included
file rather than the body file. Add support for it in
Instrument.Common.Append_SCO.

This also enables the removal of the code that was filling the
SCO_Unit_Table both in Instrument.{C,Ada_Unit}.

Change-Id: I28f7a0a1507a0ba658317b133822fb065a063ce7
TN: V728-002