setup_rts: Do not require a runtime project when Ada is disabled
Gnatcov required a runtime to be available when loading
gnatcov_rts.gpr in order to be able to choose an appropriate RTS
profile. This is not possible when Ada is not in the list of
available languages.

This commit lifts the runtime requirement and adds a fallback
strategy for when Ada is unavailable, consisting in selecting a
full runtime for the native case, and an embedded runtime for
cross targets.

Change-Id: I3eacf3eee4341565e9ece2ee9eff3ba34d2ce9b7
TN: V913-008