Instrument: allow one set of coverage buffers per source file
Now that GNATcov_RTS supports it, extend the instrumentation
infrastructure to allow it: stop assuming that there is only one set of
coverage buffers in common code, and extend data structures in the C/C++
instrumenter to be ready for creating buffers for C headers.

This is still preparatory work: just refactoring, no behavior change
expected.

Change-Id: I247fae8d3803412fc5941a6993de9709063d74fa
TN: V728-002