C instrumentation: fix parsing of preprocessed file
When parsing a preprocessed file, we were passing all of the macro
definition flags, even if the file already had been preprocessed. This
meant that self-referencing macros, that expand to themselves (e.g.
-DX=<stmt>;x), were expanded twice.

This could result in spurious coverage expectations (<stmt>;<stmt>;x with
the above example), as the parsed AST could have more coverage
obligations than the compiled code.

In practice, this meant warnings at gnatcov instrument time, with
mismatching coverage obligations between the preprocessing info recording
pass, and the instrumentation pass, and also overlapping SCOs (because we
ended up with macro locations in the instrumentation pass, which is not
supposed to happen), leading us to discard valid SCOs.

TN: V914-021