testsuite: detect more error/warnings from "gnatcov instrument"
Consider error messages (starting with !!!), and other messages starting
with "warning:" as warning messages. Note that not all error messages
lead to a fatal error, so some warnings/errors are currently ignored
even when register_failure=True.

As we still want to detect such messages (as they are a sign of
something not correct in the gnatcov execution), signal them when
warnings_as_errors=True is passed to xcov_instrument.

TN: V922-008
Depends-On: I3c7ac2bc9fc634b4769854dce1de23e1fffc5b8d
Change-Id: I9e86e1c91017e2ba98bd28325b4ef70389c9826c