Instrument.C: do not insert an atexit declaration if not needed
With --dump-trigger=atexit, gnatcov instrument inserts a all to atexit
and generates an extern declaration for it, in case it was not included.
As observed, the inserted extern declaration can conflict with the
atexit declaration coming from the inclusion of stdlib.h, for instance,
as the signature may differ.

To prevent that from happening, we will only emit a declaration of
atexit if it is not already declared in the translation unit.

Change-Id: Ie6ce9e5115d196e009ee47bbd954889c4d12f43a
TN: V914-021