Instrument.C: fix dump-trigger when main is in linkage specifier
If the main was in a linkage specifier section, it was not detected and
so gnatcov instrument did not insert any code to dump buffers. Fix the
detection of the main in that case.

TN: V916-012
Depends-On: Ib1d58309772a03453e317d42ce06d76b60df5e5b
Change-Id: I3335a8333ebdc34bf1e420b63fe981c6109d546f
(cherry picked from commit 7880c3c9b341d3fe8fbba07990b71eba67c220fa)