setup_rts: fix the loading of projects installed in the toolchain prefix
Trying to resolve the project file name with GPR2.Project.Create is
wrong as this only looks up directories listed in the GPR_PROJECT_PATH
environment variable: loading projects is also supposed to use the
target information to look up projects in the toolchain installation
prefix.

As a consequence, "gnatcov instrument" fails to load the setup
configuration when it is supposed to load gnatcov_rts.gpr as installed
by "gnatcov setup" (no --prefix): GPR2 then tries to load
gnatcov_rts.gpr in the current directory, which fails.

Directly create a GPR2.Path_Name object instead with no resolution, and
let project loading do the resolution.

Note that our testing infrastructure does not allow us to install
projects in the toolchain prefix, so it is unfortunately not possible to
add an automatic regression test for this.

TN: V921-005
Change-Id: I1e720177fded95c004fcbaff9bcc5fae23ea8a14