Instrument: handle MCDC "path overflow" for C/C++
In Ada, instrumentation is now able to skip instrumenting an MC/DC
decision if it is too complex (number of paths too high) and properly
report this to users. Similar support is missing in the C/C++
instrumenter.

This commit adds it back, extracting common code from the Ada and C/C++
instrumenter to the Instrumenter.Common unit, and finally updates the
warning message wrt recent developments (undetermined coverage status).

TN: VA04-026
Change-Id: I016ce53a827dd96c4779cfc587f60f3e9a603a21
Depends-On: Id47c316aacbc8c067ff75a2a703cf07bef5c450d