SC_Obligations: fix loading of non-instrumented SCOs from checkpoints
Move the code that loads the list of non-instrumented SCOs for a CU_Info
before CU_Info.First_SCO and CU_Info.Last_SCO are relocated, as these
are used as bound guards for the non-instrumented SCO that was just
loaded from the checkpoint (and which thus was not relocated).

For instance, suppose that we load the following CU_Info from a
checkpoint:

  .First_SCO = 10
  .Last_SCO = 20

and that after relocation, we have:

  .First_SCO = 110
  .Last_SCO = 120

(because there are already 100 SCOs in gnatcov's internal tables)

Then trying to load the non-instrumented SCO #15 will be discarded, as
15 is in range before relocation (10 .. 20), but is not in range anymore
after relocation (110 .. 120).

TN: VA04-023
Change-Id: I32c5958103bd6bc2bd26235e46b6e624073695db