Instrument: fix handling of mains given on the cmdline for units of interest
Mains specified only on the command line that were already units of
interest were instrumented as any other source file. Now they are marked
as mains and can be instrumented as such.

Also, when adding a unit of interest, always mark it as not being a main.
The decision according to the gpr file and the command line is made later
so no such computation needs to be done here.

TN: V911-003
Change-Id: Ica7c973433257fc4c62b83887efcb36cb6d29040