Instrument.C: rework "sources of interest" filtering
This is preparatory work to support the coverage of C headers: no change
of behavior expected for now.

This commit extends code that determines if a Clang cursor belongs to a
source file for which we want to compute code coverage (ad hoc notion of
"source of interest", previously called "unit of interest", but C
headers are not units of interest on their own for gnatcov). Instead of
only computing if it is a source of interest, it now also registers it
in the gnatcov's file table and computes its related project file.

Both bits of information will be necessary when creating dedicated
coverage buffers for C headers.

TN: V728-002
Change-Id: I25b65c4a4a1887036b52babc7fa1598742091f69