Instrument: make the context CU component specific to Ada
This component assumes the the instrumenter deals with exactly one
source file at a time. While this is true for Ada, this will no longer
be the case for C/C++ once we start computing code coverage for C
headers. Move the CU component from common code to Ada-specific code to
reflect that.

TN: V728-002
Change-Id: I5b3d4bff50bac15a856e1fa8ed1291d089dbd813
Depends-On: I90ce80bb9d5fb138228a8a3dcc500fc7428d184c