Add new Ada main autodump instrumentation schemes
The current autodump instrumentation scheme introduces a nested
scope in the main, containing the original main's declarations
and statements, which is incompatible with the presence of
subprogram stubs in the main.

This change modifies the autodump instrumentation shceme as
follows:
 - For atexit and ravenscar-task-termination, the procedure
   registering the buffer dump callback is now transformed into
   a function, which is used to initialize a dummy variable
   declaration inserted as first declaration in the main. This
   removes the need for a nested scope for those two dump methods.

 - For main-end, there are now two instrumentation schemes
   depending on the capabilities of the runtime.

   If finalization is available in the configured runtime, the
   buffer dump procedure is called by the finalization of a
   controlled object declared first in the main's declarative
   part.

   If the runtime doesn't provide support for finalization, then
   calls to the Dump_Buffer procedure are inserted at each exit
   point of the main (last statement of the top-level handeled
   statements, last statement of each exception handler branch and
   before each return statement).

In the case where finalization is not supported by the runtime, but
tasks are, there is no main-end dump method that works with all
constructs and still permits coverage of code in tasks depending on
the main. In such circumstances, the selected dump trigger is changed
to ravenscar-task-termination.

The feature availability checks for the runtime are implemented by
checking the presence of the required units for each dump trigger
in the runtime sources. We additionnaly check that there aren't any
configuration pragmas restricting the use of finalization when
instrumenting the main.

Change-Id: Id1f8737b737a26b1cef5e825eb3f94914afdb96a
TN: V905-005
(cherry picked from commit 4af84a4fe6f9b04ed7bfaa10c9daae8810a0760b)