Use config pragmas when processing non short-circuit operators
With this change, the configuration pragma Short_Circuit_And_Or
is taken into account to determine if gnatcov should consider non
short circuit boolean operators as short circuiting ones. The switch
already in place to enable this behavior remains in place as it would
allow a user to get MC/DC coverage on non-short circuit conditions
without modifying the semantics of the code.

TN: U621-005