Instrumentation.C: pass included project source dirs as -I args
Before, only the source directories of the project being compiled were
passed as arguments to the -I options. This poses a problem for sources
including headers from a project dependency (with "proj.gpr") as the
source directories of the included projects are not registered to be
searched during preprocessing, leading to the source files not being
found. Now all source directories including those of included projects
are given to -I.

TN: VB03-024
Change-Id: I6f2bd6e597c6f96a09026bc939c37692fd6adece