C instrumentation: change curlify implementation
When instrumenting statement expressions that are inside a control flow
statement, we may need to insert braces to preserve the semantics of the
program when inserting witness code.

We were not doing it in a robust enough way: as the end sloc of a
statement expression in clang refers to the end of the expression, and
not to the terminating semicolon, we were inserting after the next token
of the statement expression to work around it. The instrumentation of
the following code excerpt:

if (a)
   return 0
   // dumb-token
   ;
else
   ...

yielded:

if (a)
   {witness;
    return 0
    // dumb-token
    };
else
   ...

which is wrong (and not compilable code).

TN: VB03-009
(cherry picked from commit 82578e4925482a57fd5265cbc4a6e2d76be154d1)