sc_obligations.adb: ignore control locations with the instrumenter
The control location is computed from the given location for the
decision SCO, when appending an entry to the SCO table. With the C
instrumenter, such location can conflicts when decisions nest.

This problem was observed on the following code excerpt:

if (<cond> ? 1 : 0) {
   <stmt>;
}

In this case, the location of the if decision, and the location of the
ternary expression decision are the same (start location of <cond>).

As we don't need control locations for the instrumenter (they are only
useful for dominance markers, which we suppressed before), we will
ignore them when processing SCOs for source traces.

TN: VB03-009