Introduce fingerprints for buffer bit maps in SIDs and source traces
This new fingerprint allows "gnatcov coverage" to detect and reject a
source trace entry that reference bit maps which do not correlate with
the bit maps that gnatcov read from SID files.

This can arise when instrumenting and running a program a first time to
create both SID files and source traces, and then re-instrumenting that
program, either with a different version of gnatcov (which can allocate
different buffer bits for the same SCOs) or with different parameters
(for instance a different coverage level, which creates different buffer
bits).

Having a fingerprint to check SID/source trace files consistency allows
gnatcov to reject the inconsistency with a helpful warning (pointing at
the source of the inconsistency), rather than crashing as it does
currently.

Note that it is not possible to use the existing CU fingerprint,
computed from low level SCOs, since we need it to be computed when
creating the CU itself (the fingerprint is used to check whether another
CU exists for the same unit). Bit maps are computed at a later stage
(only once the high level SCOs have been created), so bit maps cannot
contribute to the CU fingerprint computation.

TN: VB21-022
Change-Id: I2254511415e0961ddcd4e8d6ed9e57facc962daf
Depends-On: I307f4283ef5e0684b8c1d6962a2960368ebe4b9d