Add default values for options --level and --annotate
This change adds default values for options --level and --annotate
to respectively stmt and report. Additionnaly, when the coverage
level is not specified on the command line or the project file,
a warning will be emitted to let the user know that the default
"stmt" value was used.

TN: VB18-035
(cherry picked from commit bcffa68c153f8d32cfa0b4317e591f1df18d9230)