Add default values for options --level and --annotate
This change adds default values for options --level and --annotate
to respectively stmt and report. Additionnaly, when the coverage
level is not specified on the command line or the project file,
a warning will be emitted to let the user know that the default
"stmt" value was used.

TN: VB18-035