Instrument.C: support SCOs in header file for C/C++ scoped metrics
The registering of the scopes is done during the instrumentation pass
after the preprocessing pass. Since the location of a scope is local to
a file, the location of SCOs defined in an included header was undestood
as being in the file being instrumented, which is incorrect.

This commit changes the storing of scopes for C/C++ by introducing a
mapping from the files to the scopes defined in them.

TN: V923-016