Instrument.C: fix main-end dump-trigger
Add a call to dump buffers at the end of the main if the last statement
is not a return statement. gnatcov used to not dump buffers in that
case, which meant that no trace was produced by the instrumented main
execution.

TN: VB28-007
(cherry picked from commit d33288380418b2432eadbc1d5d912c9c16f791e4)