Instrument.C: fix text insertion at the end of a function
gnatcov may need to insert text right before the end of the function.
The implementation was not robust enough as shown by a recent
regression.

Add a new API CX_Rewriter_Insert_Text_Before_Token to be able to insert
text right before the function closing brace, and use it instead of
trying to insert after the last statement, or after the function closing
brace (when adding new closing braces).

TN: VB28-007