Instrumentation: refactor cleaning of gnatcov-instr directory
This is preliminary work for the parallelized instrumentation.

To avoid leaving stale instrumented files (if a file instrumented in a
previous run is not a unit of interest anymore, and thus is not
instrumented anymore), we need to clean the <prj>-gnatcov-instr
directories when running gnatcov instrument.

We used to cleaning after having generated the instrumented files: we
recorded what files had been generated, and cleaned all of the files
that weren't generated by the current run.

To simplify the process, and avoid relying on the source instrumentation
communicating to the project instrumentation the sources it created, we
will clean up beforehand all of the <prj>-gnatcov-instr directories in
the project tree.

Ref: eng/cov/gnatcoverage#18