Instrument: refactor stubs
In some cases, gnatcov needs to be built w/o support for C/C++
instrumentation (for the community edition), and even w/o Ada
instrumentation support (when building the gnatcov32 driver that is not
used for instrumentation purposes).

The community edition used to crash on projects containing a mix of Ada
and C sources, when C instrumentation was enabled. It was not a problem
for editions prior to 23.0, as C was not enabled by default. Now that C
is activated in the default configuration, we need to rework this part.

Also simplify the stubbing mechanism to make it easy to deactivate Ada
instrumentation, C instrumentation or both.

TN: VC07-019