Rename C qualification test to shorten test names
The mains in the test
Qualif/C/stmt/Core/08_Statements/05_Iteration/02_DoWhile had
relatively long names, which were duplicated when testing, as each
test takes place in a subdirectory named after the main, resulting
in paths over 260 characters in length and thus being unusable on
Windows. This change shortens the names of the test mains by using
initials for each of the words that constituted the filename.

Fixes eng/cov/gnatcoverage#8