Handle C compilation units starting with a symbol / upper-case letter
The name of the slug for e.g. Pkg.c belonging to project prj was
prj_z_z__4bkg_2ec, resulting in a coverage buffer symbol name with two
underlines. This is not accepted by the Ada compiler, so it resulted in a
compile error as far as projects mixing Ada / C code are concerned.

To fix this, we change the slug computation mechanism to avoid generating
a slug with consecutive underlines.

TN: VB03-009