Implement first two levels of assertion coverage
This change adds the possibility of computing the coverage of assertions
and contracts for levels:

* ATC (Assertion True Coverage): an assertion is covered if its whole
decision has been evaluated to True at least once.
* ATCC (Assertion True Condition Coverage): and assertion is covered if
every condition of the decision has been evaluated at least once in the
context of an evaluation to True of the whole decision.

The relevant assertions and contracts are Assert, Assert_And_Cut,
Assume, Check, Loop_Invariant, Type_Invariant, Precondition,
Postcondition.

The only unsupported cases are Pre/Postcondition and Type_Invariant as
pragma statements which are not instrumented. A warning is issued
stating as much, and the rest of the code is instrumented as normal.

Issue: #7
TN: O212-062