Remove obsolete ref to bin traces in consolidation doc
The statement telling that mixing bin and src traces isn't
supported is in the "introduction" chapter of the bin traces
appendix. There's no need to have it in the main chapter on
consolidation, and it actually is confusing there.

Part of V805-004

(cherry picked from commit a4bcf434904c7991b20406a13e54882b729db76f)