Remove a few src/bin traces distinctions
Simply refer to "traces" instead of "source or binary traces"
in a few places, where the explicit mention of the two isn't of
importance and is potentially confusing in a context where
binary traces have not been introduced.

Simplify the section describing handling of source relocation,
removing non fundamental detail on how default locations are picked
depending on the trace kind.

Part of V805-004