SC_Obligations: Fix consolidation bug for exemption annotations
When loading an exemption annotation from a checkpoint, the CU to
which the annotation belongs is not updated according to the mapping
of CU between the checkpoint and gnatcov internal tables. This is
not a problem during the current execution of gnatcov as the CU of
exemption annotations are only used when loading checkpoints, but
if the annotation with the wrong CU is then written back to a
checkpoint, then this can lead to either annotations wrongly not
being loaded, or a constraint error due to attempts to insert an
annotation with a key already present in the annotation map.

See the corresponding test for a concrete example of the above.

This change fixes this bug by inserting the annotations with the CU
correctly remapped.

Note that this bug was present from the introduction of checkpoints,
so it potentially concerns all previous gnatcov versions.

TN: VC14-043