Clarify Makefile/gnatcov.gpr split of responsibilities
The gnatcov.gpr project file is used to build a variety
of executables, with some parameters controlling the possible
stubbing of some units.

The control of what sources get stubbed was split between
Makefiles and the project file, with Makefiles adjusting
settings based on knowledge of what the project file does
e.g. for 32 vs 64bit.

As we are going to introduce support for 32 vs 64bit
variants of the trace32 adapter, which also rely on gnatcov.gpr,
some implicit assumptions in the project file won't hold any
more and getting the current split to work for all combinations
induces undue conceptual complexity.

This change simplifies the overall organization by removing
some attempts at figuring default parameter values from the
project file, pushing the responsibility to do so entirely
in Makefiles instead.

To let Makefiles associate meaningful parameter values for the
various uses of the project file, the change also removes the
computation of a Main attribute from the project itself.
Main units are simply stated from build commands when they
are needed, together the parameters that make sense for what
is being built.

Essentially, this defers the responsibilty to state what is
needed at the points where the build commands are issued, which
is where we know best what we intend to build and where what is
needed is simple to state locally.

Part of W105-035