Trace32: Also build the adapter for 64 bit tragets
The Trace32 adapter was linked against the 32 version of gnatcov,
and could thus only process traces from a 32bit target.

This change adds a second executable for the adapter, linked
against the 64 bit version of gnatcov to allow processing of
traces from 64 bit targets. This also changes the adapter
invocation to ensure the correct adapter is used depending on the
target.

Part of W105-035