Qualif test for unit selection attributes in non root project
Our qualification tests for the use of project files
to select units of interest had essentially

- a set of tests showing proper selection of the projects
  of interest, from command line switches,

- a set of tests showing proper selection of units of interest
  in a project, from attributes in a project file.

The use of attributes in project files was in effect only
tested within a root project.

This change adds a test to check that the attributes also
work within a non root project.

Part of T317-055 (qualification material for use of project files)