Instrument.C: fix name used for the top level scope
We can't rely on the Get_Unique_Name API as it freezes the files table
and forbids insertions of filenames afterwards. Using the simple name
here is unambiguous, as scopes are displayed in index-based reports
(xml and dhtml), in each file report.

The issue was never visible in production, as assertion are
deactivated. It was not visible in dev as well as we probably don't
have tests running instrumentation over multiple C files.

TN: V923-016